chore: move polling to async query service
